Extracted into TaskVerifier class
